backend/requirements.txt
fastapi
uvicorn
python-multipart
python-jose
passlib
bcrypt
PyJWT
pydantic
python-dotenv==1.0.0
Pillow
PyPDF2
pytesseract
openai
pathlib